tpmclient: Refactor SimpleHmacOrPolicyTest into separate integration test.
This test uses an NV region to test both HMAC and POLICY sessions. This
commit removes the SimpleHmacOrPolicyTest from tpmclient and moves it to
test/integration/sapi-hmac-auth.int.c. The code has also been refactored
to remove the need for the boolean flag that determined whether or not
the session is setup as an HMAC or POLICY session. The policy is now
established in a separate function and passed as a parameter (dependency
inversion).

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>